requirements.txt
flask
requests
slack
slackeventsapi
python-dotenv
